+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Changed TriangleMesh to store materials in a list so they can be accessed by the material index (PR #5938)
 * Support multi-threading in the RayCastingScene function to commit scene (PR #6051).
 * Fix some bad triangle generation in TriangleMesh::SimplifyQuadricDecimation
+* Check for support of CUDA Memory Pools at runtime (#4679)
 
 ## 0.13